Implement PaC pruning on COmponent deletion
